+++ src/shared/client/api.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: AGPL-3.0-only
 # SPDX-FileCopyrightText: 2024 Univention GmbH
 import logging
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Callable, Coroutine, Dict, List, Optional, Tuple
 
 import aiohttp
 
@@ -11,7 +11,10 @@ from shared.models import (
     Event,
     ProvisioningMessage,
     MessageProcessingStatusReport,
+    MessageProcessingStatus,
 )
+
+from shared.models.queue import Message
 
 from shared.client.config import settings
 
@@ -32,8 +35,7 @@ class AsyncClient:
         )
 
         async with aiohttp.ClientSession(raise_for_status=True) as session:
-            # TODO: do this with propper logging
-            print(subscription.model_dump())
+            logger.debug(subscription.model_dump())
             async with session.post(
                 f"{settings.consumer_registration_url}/subscriptions",
                 json=subscription.model_dump(),
@@ -114,3 +116,95 @@ class AsyncClient:
             ):
                 # either return nothing or let `.post` throw
                 pass
+
+
+class MessageHandler:
+    def __init__(
+        self,
+        client: AsyncClient,
+        subscription_name: str,
+        callbacks: List[Callable[[Message], Coroutine[None, None, None]]],
+        pop_after_handling: bool = True,
+        message_limit: Optional[int] = None,
+    ):
+        """
+        Each callback should be an asynchronous function to facilitate downstream asynchronous operations.
+        To prevent race conditions between message processing and callback execution,
+        each message and its callbacks are handled sequentially.
+
+        Args:
+            client: AsyncClient
+            subscription_name: The name of the target subscription to listen on.
+            callbacks: A list of asynchronous callback functions, each accepting a single message object as a parameter.
+            message_limit: An optional integer specifying the maximum number of messages to process,
+                primarily to facilitate testing.
+            pop_after_handling: If False, messages are acknowledged immediately upon reception
+                rather than after all callbacks for the message have been successfully executed.
+        """
+        if not callbacks:
+            raise ValueError("Callback functions can't be empty")
+        self.client = client
+        self.subscription_name = subscription_name
+        self.callbacks = callbacks
+        self.pop_after_handling = pop_after_handling
+        self.message_limit = message_limit
+
+    async def _callback_wrapper(
+        self,
+        message: ProvisioningMessage,
+        callback: Callable[[Message], Coroutine[None, None, None]],
+    ):
+        """
+        Wrapper around a client's callback function to encapsulate
+        error handling and message acknowledgement
+        """
+        # TODO: Message is not enough, we need a specific ClientMessage, that includes for example num_redelivered
+        await callback(Message.model_validate(message.data))
+        if not self.pop_after_handling:
+            return
+        try:
+            # TODO: Retry acknowledgement
+            report = MessageProcessingStatusReport(
+                status=MessageProcessingStatus.ok,
+                message_seq_num=message.sequence_number,
+                publisher_name=message.publisher_name,
+            )
+            await self.client.set_message_status(self.subscription_name, [report])
+        except (
+            aiohttp.ClientError,
+            aiohttp.ClientConnectionError,
+            aiohttp.ClientResponseError,
+        ) as exc:
+            # TODO: Test this
+            logger.error(
+                "Failed to acknowledge message. It will be redelivered later...",
+                exc.message,
+            )
+
+    async def run(
+        self,
+    ):
+        """
+        This method is designed for asynchronous execution, either by awaiting it within an async context
+        or using `asyncio.run()`.
+        It continuously listens for messages, either indefinitely or until a specified message limit is reached, and
+        invokes a series of callbacks for each message.
+        """
+        counter = 0
+
+        while True:
+            messages = await self.client.get_subscription_messages(
+                self.subscription_name,
+                count=1,
+                timeout=10,
+                # TODO: pop is broken serverside at the moment
+                # pop= not pop_after_handling,
+            )
+
+            for message in messages:
+                for callback in self.callbacks:
+                    await self._callback_wrapper(message, callback)
+                if self.message_limit:
+                    counter += 1
+                    if counter >= self.message_limit:
+                        return