+++ tests/e2e/test_consumer_client.py
@@ -1,43 +1,17 @@
 # SPDX-License-Identifier: AGPL-3.0-only
 # SPDX-FileCopyrightText: 2024 Univention GmbH
 
-import uuid
 import pytest
-import requests
 
 import shared.client
-from shared.client.config import settings
-from shared.models.api import MessageProcessingStatus
 import shared.models.queue
 
-from tests.conftest import (
-    REALM,
-    TOPIC,
-    REALMS_TOPICS,
+from tests.e2e.helpers import (
+    create_message_via_events_api,
+    create_message_via_udm_rest_api,
+    pop_all_messages,
 )
-from shared.models import PublisherName
 from univention.admin.rest.client import UDM
-
-settings.provisioning_api_host = "localhost"
-
-
-@pytest.fixture
-def provisioning_client() -> shared.client.AsyncClient:
-    return shared.client.AsyncClient()
-
-
-@pytest.fixture
-async def simple_subscription(provisioning_client: shared.client.AsyncClient):
-    subscriber_name = str(uuid.uuid4())
-    await provisioning_client.create_subscription(
-        name=subscriber_name,
-        realms_topics=REALMS_TOPICS,
-        request_prefill=False,
-    )
-
-    yield subscriber_name
-
-    await provisioning_client.cancel_subscription(subscriber_name)
 
 
 async def test_create_subscription(
@@ -60,27 +34,20 @@ async def test_get_empty_messages(
 
 
 async def test_send_message(
-    provisioning_client: shared.client.AsyncClient, simple_subscription: str
+    provisioning_client: shared.client.AsyncClient,
+    simple_subscription: str,
+    provisioning_base_url: str,
 ):
-    requests.post(
-        "http://localhost:7777/events/v1/events",
-        json={
-            "publisher_name": PublisherName.consumer_client,
-            "ts": "2024-02-07T09:01:33.835Z",
-            "realm": REALM,
-            "topic": TOPIC,
-            "body": {"foo": "bar"},
-        },
-    )
+    data = create_message_via_events_api(provisioning_base_url)
 
     response = await provisioning_client.get_subscription_messages(
         name=simple_subscription,
         count=1,
-        timeout=1,
+        timeout=10,
     )
 
     assert len(response) == 1
-    assert response[0].body["foo"] == "bar"
+    assert response[0].data["body"] == data
 
 
 @pytest.mark.xfail()
@@ -97,11 +64,7 @@ async def test_pop_message(
 async def test_get_real_messages(
     provisioning_client: shared.client.AsyncClient, simple_subscription: str, udm: UDM
 ):
-    groups = udm.get(TOPIC)
-    assert groups
-    group = groups.new()
-    group.properties["name"] = str(uuid.uuid1())
-    group.save()
+    group = create_message_via_udm_rest_api(udm)  # noqa: F841
 
     response = await provisioning_client.get_subscription_messages(
         name=simple_subscription,
@@ -109,6 +72,17 @@ async def test_get_real_messages(
     )
 
     assert len(response) == 1
+
+
+async def test_get_multiple_messages(
+    provisioning_client: shared.client.AsyncClient, simple_subscription: str, udm: UDM
+):
+    group1 = create_message_via_udm_rest_api(udm)  # noqa: F841
+    group2 = create_message_via_udm_rest_api(udm)  # noqa: F841
+    group3 = create_message_via_udm_rest_api(udm)  # noqa: F841
+
+    result = await pop_all_messages(provisioning_client, simple_subscription, 4)
+    assert len(result) == 3
 
 
 @pytest.mark.xfail()
@@ -121,27 +95,3 @@ async def test_get_messages_zero_timeout(
     )
 
     assert response == []
-
-
-async def test_acknowledge_messages(
-    provisioning_client: shared.client.AsyncClient, simple_subscription: str, udm: UDM
-):
-    groups = udm.get(TOPIC)
-    assert groups
-    group = groups.new()
-    group.properties["name"] = str(uuid.uuid1())
-    group.save()
-
-    response = await provisioning_client.get_subscription_messages(
-        name=simple_subscription,
-        timeout=5,
-    )
-
-    assert len(response) == 1
-    message = response[0]
-
-    response = await provisioning_client.set_message_status(
-        simple_subscription,
-        message,
-        MessageProcessingStatus.ok,
-    )